requirements.txt
torch>=2.0.0
numpy>=1.21.0
Pillow>=9.0.0
matplotlib>=3.5.0
tqdm>=4.64.0
opencv-python
Flask>=2.0.0
Flask-Cors
